// roseedhermon-ui-ng/src/app/api/model/member.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Member { 
    id?: string;
    lastName?: string;
    firstName?: string;
    gender?: Member.GenderEnum;
    birthDate?: string;
    profession?: string;
    phoneNumber?: string;
    email?: string;
    address?: string;
    city?: string;
    location?: string;
    photo?: string;
    socialLinks?: Array<string>;
}
export namespace Member {
    export type GenderEnum = 'MALE' | 'FEMALE' | 'OTHER';
    export const GenderEnum = {
        Male: 'MALE' as GenderEnum,
        Female: 'FEMALE' as GenderEnum,
        Other: 'OTHER' as GenderEnum
    };
}